// src/app/api/model/tvSeriesDetails200Response.ts
/**
 * tmdb-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TvSeriesDetails200ResponseNetworksInner } from './tvSeriesDetails200ResponseNetworksInner';
import { TvSeriesDetails200ResponseCreatedByInner } from './tvSeriesDetails200ResponseCreatedByInner';
import { TvSeriesDetails200ResponseGenresInner } from './tvSeriesDetails200ResponseGenresInner';
import { TvSeriesDetails200ResponseProductionCountriesInner } from './tvSeriesDetails200ResponseProductionCountriesInner';
import { TvSeriesDetails200ResponseProductionCompaniesInner } from './tvSeriesDetails200ResponseProductionCompaniesInner';
import { TvSeriesDetails200ResponseLastEpisodeToAir } from './tvSeriesDetails200ResponseLastEpisodeToAir';
import { TvSeriesDetails200ResponseSeasonsInner } from './tvSeriesDetails200ResponseSeasonsInner';
import { MovieDetails200ResponseSpokenLanguagesInner } from './movieDetails200ResponseSpokenLanguagesInner';


export interface TvSeriesDetails200Response { 
    adult?: boolean;
    backdrop_path?: string;
    created_by?: Array<TvSeriesDetails200ResponseCreatedByInner>;
    episode_run_time?: Array<number>;
    first_air_date?: string;
    genres?: Array<TvSeriesDetails200ResponseGenresInner>;
    homepage?: string;
    id?: number;
    in_production?: boolean;
    languages?: Array<string>;
    last_air_date?: string;
    last_episode_to_air?: TvSeriesDetails200ResponseLastEpisodeToAir;
    name?: string;
    next_episode_to_air?: any | null;
    networks?: Array<TvSeriesDetails200ResponseNetworksInner>;
    number_of_episodes?: number;
    number_of_seasons?: number;
    origin_country?: Array<string>;
    original_language?: string;
    original_name?: string;
    overview?: string;
    popularity?: number;
    poster_path?: string;
    production_companies?: Array<TvSeriesDetails200ResponseProductionCompaniesInner>;
    production_countries?: Array<TvSeriesDetails200ResponseProductionCountriesInner>;
    seasons?: Array<TvSeriesDetails200ResponseSeasonsInner>;
    spoken_languages?: Array<MovieDetails200ResponseSpokenLanguagesInner>;
    status?: string;
    tagline?: string;
    type?: string;
    vote_average?: number;
    vote_count?: number;
}