// src/app/api/model/tvSeasonDetails200Response.ts
/**
 * tmdb-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TvSeriesDetails200ResponseNetworksInner } from './tvSeriesDetails200ResponseNetworksInner';
import { TvSeasonDetails200ResponseEpisodesInner } from './tvSeasonDetails200ResponseEpisodesInner';


export interface TvSeasonDetails200Response { 
    _id?: string;
    air_date?: string;
    episodes?: Array<TvSeasonDetails200ResponseEpisodesInner>;
    name?: string;
    networks?: Array<TvSeriesDetails200ResponseNetworksInner>;
    overview?: string;
    id?: number;
    poster_path?: string;
    season_number?: number;
    vote_average?: number;
}